requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
python-dotenv==1.0.0
pydantic==2.6.1
httpx==0.24.1
supabase==2.3.5 